Comments, extraction, modifiers comments addressed.
