+++ community/lucene-index/src/test/java/org/neo4j/index/impl/lucene/ReadOnlyIndexReferenceTest.java
@@ -19,17 +19,6 @@
  */
 package org.neo4j.index.impl.lucene;
 
-import org.apache.lucene.document.Document;
-import org.apache.lucene.document.FieldSelector;
-import org.apache.lucene.index.CorruptIndexException;
-import org.apache.lucene.index.FieldInfos;
-import org.apache.lucene.index.IndexReader;
-import org.apache.lucene.index.Term;
-import org.apache.lucene.index.TermDocs;
-import org.apache.lucene.index.TermEnum;
-import org.apache.lucene.index.TermFreqVector;
-import org.apache.lucene.index.TermPositions;
-import org.apache.lucene.index.TermVectorMapper;
 import org.apache.lucene.search.IndexSearcher;
 import org.junit.Before;
 import org.junit.Rule;
@@ -37,7 +26,6 @@ import org.junit.Test;
 import org.junit.rules.ExpectedException;
 
 import java.io.IOException;
-import java.util.Map;
 
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
@@ -52,7 +40,7 @@ public class ReadOnlyIndexReferenceTest
     public ExpectedException expectedException = ExpectedException.none();
     private IndexIdentifier identifier = mock( IndexIdentifier.class );
     private IndexSearcher searcher = mock( IndexSearcher.class );
-    private TestIndexReader reader = new TestIndexReader();
+    private CloseTrackingIndexReader reader = new CloseTrackingIndexReader();
     private ReadOnlyIndexReference indexReference = new ReadOnlyIndexReference( identifier, searcher );
 
     @Before
@@ -161,147 +149,4 @@ public class ReadOnlyIndexReferenceTest
     {
         assertTrue( indexReference.close() );
     }
-
-    private class TestIndexReader extends IndexReader
-    {
-        private boolean closed = false;
-
-        @Override
-        public TermFreqVector[] getTermFreqVectors( int docNumber ) throws IOException
-        {
-            return new TermFreqVector[0];
-        }
-
-        @Override
-        public TermFreqVector getTermFreqVector( int docNumber, String field ) throws IOException
-        {
-            return null;
-        }
-
-        @Override
-        public void getTermFreqVector( int docNumber, String field, TermVectorMapper mapper ) throws IOException
-        {
-
-        }
-
-        @Override
-        public void getTermFreqVector( int docNumber, TermVectorMapper mapper ) throws IOException
-        {
-
-        }
-
-        @Override
-        public int numDocs()
-        {
-            return 0;
-        }
-
-        @Override
-        public int maxDoc()
-        {
-            return 0;
-        }
-
-        @Override
-        public Document document( int n, FieldSelector fieldSelector ) throws CorruptIndexException, IOException
-        {
-            return null;
-        }
-
-        @Override
-        public boolean isDeleted( int n )
-        {
-            return false;
-        }
-
-        @Override
-        public boolean hasDeletions()
-        {
-            return false;
-        }
-
-        @Override
-        public byte[] norms( String field ) throws IOException
-        {
-            return new byte[0];
-        }
-
-        @Override
-        public void norms( String field, byte[] bytes, int offset ) throws IOException
-        {
-
-        }
-
-        @Override
-        protected void doSetNorm( int doc, String field, byte value ) throws CorruptIndexException, IOException
-        {
-
-        }
-
-        @Override
-        public TermEnum terms() throws IOException
-        {
-            return null;
-        }
-
-        @Override
-        public TermEnum terms( Term t ) throws IOException
-        {
-            return null;
-        }
-
-        @Override
-        public int docFreq( Term t ) throws IOException
-        {
-            return 0;
-        }
-
-        @Override
-        public TermDocs termDocs() throws IOException
-        {
-            return null;
-        }
-
-        @Override
-        public TermPositions termPositions() throws IOException
-        {
-            return null;
-        }
-
-        @Override
-        protected void doDelete( int docNum ) throws CorruptIndexException, IOException
-        {
-
-        }
-
-        @Override
-        protected void doUndeleteAll() throws CorruptIndexException, IOException
-        {
-
-        }
-
-        @Override
-        protected void doCommit( Map<String,String> commitUserData ) throws IOException
-        {
-
-        }
-
-        @Override
-        protected void doClose() throws IOException
-        {
-            closed = true;
-        }
-
-        @Override
-        public FieldInfos getFieldInfos()
-        {
-            return null;
-        }
-
-        public boolean isClosed()
-        {
-            return closed;
-        }
-    }
-
 }
