+++ community/lucene-index/src/test/java/org/neo4j/kernel/api/impl/index/LuceneLabelScanStoreTest.java
@@ -132,6 +132,7 @@ public class LuceneLabelScanStoreTest
     {
         expectedException.expectCause( Matchers.<Throwable>instanceOf( IOException.class ) );
         startLabelScanStore( MapUtil.stringMap(GraphDatabaseSettings.read_only.name(), "true") );
+        assertTrue( monitor.noIndexCalled );
     }
 
     @Test
